# cmd/laas/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service to host license information for other services to query over REST API.",
        "title": "laas (License as a Service) API",
        "contact": {
            "name": "FOSSology",
            "url": "https://fossology.org",
            "email": "fossology@fossology.org"
        },
        "license": {
            "name": "GPL-2.0-only",
            "url": "https://github.com/fossology/LicenseDb/blob/main/LICENSE"
        },
        "version": "0.0.9"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/hard-reset": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete all data from configured tables used by license/obligation/audit domain",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Hard reset database",
                "operationId": "HardResetDatabase",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "failed to hard reset database",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LicenseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "path": {
                    "type": "string",
                    "example": "/api/v1/licenses"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-12-01T10:00:51+05:30"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token from /login endpoint. Enter the token with the `Bearer ` prefix, e.g. \\\"Bearer eyJhbGciOiJ.....\\\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}